[Parser] Implemented OCP for TaskCommandsParser
