+++ src/main/java/seedu/address/logic/parser/AddCommandParser.java
@@ -1,0 +1,24 @@
+package seedu.address.logic.parser;
+
+import seedu.address.commons.exceptions.IllegalValueException;
+import seedu.address.logic.commands.taskcommands.AddTaskCommand;
+import seedu.address.logic.commands.taskcommands.IncorrectTaskCommand;
+import seedu.address.logic.commands.taskcommands.TaskCommand;
+
+public class AddCommandParser extends CommandParser{
+	/**
+     * Parses arguments in the context of the add task command.
+     *
+     * @param args full command args string
+     * @return the prepared command
+     */
+	public TaskCommand prepareCommand(String arguments) {
+		// Assume arguments === description for now
+		try {
+			return new AddTaskCommand(arguments);
+		} catch (IllegalValueException ive) {
+            return new IncorrectTaskCommand(ive.getMessage());
+        }
+	
+	}
+}
