+++ src/main/java/seedu/address/logic/parser/CommandParser.java
@@ -1,0 +1,9 @@
+package seedu.address.logic.parser;
+
+import seedu.address.logic.commands.taskcommands.TaskCommand;
+
+public abstract class CommandParser {
+	
+	public abstract TaskCommand prepareCommand(String arguments);
+
+}
