+++ src/main/java/seedu/address/logic/parser/TaskCommandsParser.java
@@ -14,14 +14,12 @@ import seedu.address.logic.commands.taskcommands.DeleteTaskCommand;
 import seedu.address.logic.commands.taskcommands.IncorrectTaskCommand;
 import seedu.address.logic.commands.taskcommands.TaskCommand;
 
-public class TaskCommandsParser {
+public class TaskCommandsParser{
 	/**
      * Used for initial separation of command word and args.
      */
     private static final Pattern BASIC_COMMAND_FORMAT = Pattern.compile("(?<commandWord>\\S+)(?<arguments>.*)");
     
-    private static final Pattern TASK_INDEX_ARGS_FORMAT = Pattern.compile("(?<targetIndex>.+)");
-
     public TaskCommandsParser() {}
     
     /**
@@ -39,68 +37,8 @@ public class TaskCommandsParser {
 
         final String commandWord = matcher.group("commandWord");
         final String arguments = matcher.group("arguments");
-        switch (commandWord) {
-
-        case AddTaskCommand.COMMAND_WORD:
-            return prepareAdd(arguments);
-
-        case DeleteTaskCommand.COMMAND_WORD:
-            return prepareDelete(arguments);
-
-        default:
-            return new IncorrectTaskCommand(MESSAGE_UNKNOWN_COMMAND);
-        }
+        CommandParser command =  ParserType.get(commandWord);
+        return command.prepareCommand(arguments);
+       
     }
-
-    /**
-     * Parses arguments in the context of the delete task command.
-     *
-     * @param args full command args string
-     * @return the prepared command
-     */
-	private TaskCommand prepareDelete(String arguments) {
-		 Optional<Integer> index = parseIndex(arguments);
-	        if(!index.isPresent()){
-	            return new IncorrectTaskCommand(
-	                    String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteTaskCommand.MESSAGE_USAGE));
-	        }
-
-	        return new DeleteTaskCommand(index.get());
-	}
-
-    /**
-     * Parses arguments in the context of the add task command.
-     *
-     * @param args full command args string
-     * @return the prepared command
-     */
-	private TaskCommand prepareAdd(String arguments) {
-		// Assume arguments === description for now
-		try {
-			return new AddTaskCommand(arguments);
-		} catch (IllegalValueException ive) {
-            return new IncorrectTaskCommand(ive.getMessage());
-        }
-		
-	}
-	
-	/**
-     * Returns the specified index in the {@code command} IF a positive unsigned integer is given as the index.
-     *   Returns an {@code Optional.empty()} otherwise.
-     */
-    private Optional<Integer> parseIndex(String command) {
-        final Matcher matcher = TASK_INDEX_ARGS_FORMAT.matcher(command.trim());
-        if (!matcher.matches()) {
-            return Optional.empty();
-        }
-
-        String index = matcher.group("targetIndex");
-        if(!StringUtil.isUnsignedInteger(index)){
-            return Optional.empty();
-        }
-        return Optional.of(Integer.parseInt(index));
-
-    }
-
-   
 }
